Policyの適用

+++ app/Http/Controllers/Admin/BookController.php
@@ -12,6 +12,7 @@ use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\View\View;
@@ -40,6 +41,9 @@ class BookController extends Controller
     // createアクション
     public function create(): View
     {
+        // BookPolicyのcreateメソッドによる認可
+        $this->authorize('create', Book::class);
+
         // カテゴリ一覧
         $categories = Category::all();
 
@@ -52,6 +56,9 @@ class BookController extends Controller
     // storeアクション
     public function store(BookPostRequest $request): RedirectResponse
     {
+        // BookPolicyのcreateメソッドによる認可
+        $this->authorize('create', Book::class);
+
         // 書籍データ登録用のオブジェクトを生成する
         $book = new Book();
 
@@ -59,6 +66,7 @@ class BookController extends Controller
         $book->category_id = $request->category_id;
         $book->title = $request->title;
         $book->price = $request->price;
+        $book->admin_id = Auth::id();
 
         DB::transaction(function () use ($book, $request) {
             // 保存
@@ -75,6 +83,12 @@ class BookController extends Controller
     // editアクション
     public function edit(Book $book): View
     {
+        // 作成者以外はアクセス不可
+        // if (Auth::user()->cannot('update', $book)) {
+        //     abort(403);
+        // }
+        $this->authorize('update', $book);
+
         // カテゴリ一覧の取得
         $categories = Category::all();
 
@@ -90,6 +104,9 @@ class BookController extends Controller
     // updateアクション
     public function update(BookPutRequest $request, Book $book): RedirectResponse
     {
+        // 作成者以外はアクセス不可
+        $this->authorize('update', $book);
+
         $book->category_id = $request->category_id;
         $book->title = $request->title;
         $book->price = $request->price;
@@ -108,6 +125,8 @@ class BookController extends Controller
     // destroy
     public function destroy(Book $book): RedirectResponse
     {
+        // 作成者以外はアクセス不可
+        $this->authorize('update', $book);
         // 削除(カスケード機能による子テーブルの削除)
         $book->delete();
         return redirect(route('book.index'))->with('message', $book->title . 'を削除しました。');
